requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.1.5
django-cors-headers==2.4.0
idna==2.8
numpy==1.16.1
Pillow==5.4.1
pytz==2018.9
requests==2.21.0
urllib3==1.24.1
